backend/requirements.txt
torch
diffusers
transformers
accelerate
safetensors
opencv-python-headless
Pillow
fastapi
uvicorn[standard]
python-multipart
